pyproject.toml
[tool.poetry]
name = "sdplc"
version = "0.1.0"
description = ""
authors = ["ulfaric <ryf0510@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymodbus = "^3.6.8"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
pyyaml = "^6.0.1"
colorlog = "^6.8.2"
akatosh = "^3.0.6"
asyncua = "^1.1.0"
numpy = "^1.26.4"
fasterapi = "^0.0.41"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
